feat: add Postgres engine alongside DuckDB-WASM
Learners can now solve problems in Postgres (PGlite) or DuckDB-WASM.
Both run entirely in-browser; no server-side execution.

Schema:
- New Dialect enum (DUCKDB | POSTGRES)
- SQLProblem.dialects Dialect[] with default [DUCKDB, POSTGRES] —
  most problems are portable; authors narrow only when the canonical
  solution uses dialect-specific syntax (JSONB, STRING_AGG, LIST_AGG).
- Migration backfills every existing problem to [DUCKDB, POSTGRES].

Engine:
- New lib/pglite.ts wraps @electric-sql/pglite (~3MB Postgres-WASM,
  lazy-imported only when the learner picks Postgres).
- lib/use-problem-db.ts is now dialect-aware. Both engines lazy-load
  via dynamic imports — pure DuckDB users never download Postgres.
- Schema replays on every dialect change (clean slate).

Workspace UI:
- DialectChip toggle on the playground footer. Static badge when only
  one dialect is allowed; segmented two-button toggle when both are.
- Last-picked dialect persists per-problem in localStorage
  (`dl:dialect:<slug>`).
- DESCRIBE fallback is dialect-aware: information_schema.columns for
  Postgres, DESCRIBE "<table>" for DuckDB. Most problems use the
  server-parsed initialTableInfos and skip this path entirely.

Admin + API + MCP:
- ProblemForm: SQL engines field with dual checkboxes; can't disable
  the only remaining dialect.
- ProblemCreateInput / ProblemUpdateInput Zod schemas accept
  dialects array.
- /api/admin/problems POST + PATCH wire dialects through.
- MCP list_problems projection adds dialects; tool description
  documents it.

Seed compat:
- DOUBLE → DOUBLE PRECISION in lib/seed-data.ts and lib/seed-data-hr.ts
  so the same DDL string runs in both engines (DuckDB accepts both;
  Postgres rejects bare DOUBLE).
- Reseeded local DB so existing SqlSchema rows have the portable form.

+++ lib/seed-data-hr.ts
@@ -13,7 +13,7 @@ CREATE TABLE employees (
 
 CREATE TABLE salaries (
     employee_id INTEGER,
-    amount DOUBLE
+    amount DOUBLE PRECISION
 );
 
 INSERT INTO departments VALUES (1, 'Engineering');
